Backport dense A* core after stride rollback

+++ src/main/java/baritone/pathing/movement/MovementCatalog.java
@@ -48,7 +48,7 @@ public final class MovementCatalog {
     return primitives;
   }
 
-  public MovementPrimitive primitive(short index) {
+  public MovementPrimitive primitive(int index) {
     return primitives[index];
   }
